Refactored code

+++ devops1.py
@@ -9,19 +9,21 @@ import webbrowser
 import subprocess
 import random
 import string
+import sys
+from os.path import exists
 from time import sleep
 
-def createInstance(keyPairFilename, userData):
-    ec2 = boto3.resource('ec2')
 
+def createInstance(keyName, imageId, userData):
+    ec2 = boto3.resource('ec2')
     print("Creating a new ec2 instance...")
 
     new_instances = ec2.create_instances(
-        ImageId='ami-0aa7d40eeae50c9a9',
+        ImageId=imageId,
         MinCount=1,
         MaxCount=1,
         InstanceType='t2.nano',
-        KeyName='devopsAwsKey',
+        KeyName=keyName,
         SecurityGroupIds=['sg-076cb4038b190be4e'],
         UserData=userData
     )
@@ -30,27 +32,25 @@ def createInstance(keyPairFilename, userData):
     # Adding a name tag to the created instance
     instance.create_tags(Tags=[{'Key':'Name', 'Value':'Web Server'}])
 
-    # Print out details of created instances
-    print("Created instances:")
-    for inst in new_instances:
-        while(inst.public_ip_address is None):
-            sleep(1)
-            print("sleeping to get the IP")
-            inst.reload()
-        print("ID: " + inst.id + " Public IP: " + inst.public_ip_address)
+    # Print out details of the created instance
+    while(instance.public_ip_address is None):
+        sleep(1)
+        print("Waiting to acquire the IP address...")
+        instance.reload()
+    print("Created instance ID: " + instance.id + ", Public IP: " + instance.public_ip_address)
 
     instance.wait_until_running()
     instance.reload()
-    print("Instance running")
+    print(f"Instance {instance.id} is running.")
 
     # Checking if all UserData commands were executed and 'fileCommandsCompleted' file is on the EC2 instance
-    ssh_command = "ssh -o StrictHostKeyChecking=no -i " + keyPairFilename + " ec2-user@" + instance.public_ip_address + " 'ls'"
+    ssh_command = "ssh -o StrictHostKeyChecking=no -i " + keyName + ".pem ec2-user@" + instance.public_ip_address + " 'ls'"
     for i in range(10):
         result = subprocess.run(ssh_command, shell=True, stdout=subprocess.PIPE)
         filesList = result.stdout.decode('utf-8').strip().split('\n')
         print("Checking if Apache server is running...")
         if ("fileCommandsCompleted" in filesList):
-            print("Apache server installed and running on " + instance.public_ip_address)
+            print("Apache server installed and running at " + instance.public_ip_address)
             break
         sleep(10)
 
@@ -59,7 +59,7 @@ def createInstance(keyPairFilename, userData):
 
 def createBucket():
     s3 = boto3.resource('s3')
-
+    print("Creating a new S3 Bucket...")
     alphaNum = string.ascii_lowercase + string.digits
     bucketName = "jbloggs-" + "".join(random.sample(alphaNum, 6))
 
@@ -75,16 +75,30 @@ def createBucket():
     }
 
     bucket_website = s3.BucketWebsite(bucketName)
-
     response = bucket_website.put(WebsiteConfiguration=website_configuration)
-
     print("S3 Bucket '" + bucketName + "' created.")
 
-    return bucketName
+    return s3.Bucket(bucketName)
 
 
 if __name__ == "__main__":
-    keyPairFilename = "devopsAwsKey.pem"
+    keyName = "devopsAwsKey"
+    imageId = 'ami-006dcf34c09e50022'
+    try:
+        keyName= sys.argv[1]
+    except IndexError:
+        pass # No argument found. Using default keyPair instead
+    
+    keyFilename = keyName + ".pem"
+    if not exists(keyFilename):
+        print(f"Key Pair file {keyName}.pem not found. Closing the program...")
+        exit()
+
+    try:
+        imageId = sys.argv[2]
+    except IndexError:
+        pass
+
     userData = """#!/bin/bash
                 #yum update -y
                 yum install httpd -y
@@ -100,48 +114,36 @@ if __name__ == "__main__":
                 touch /home/ec2-user/fileCommandsCompleted"""
 
     # Creating EC2 instance
-    instance = createInstance(keyPairFilename, userData)
+    instance = createInstance(keyName, imageId, userData)
 
     # Opening the Apache test page in the browser
     webbrowser.open_new_tab('http://' + instance.public_ip_address)
 
     # Creating S3 bucket
-    bucketName = createBucket()
+    bucket = createBucket()
 
     # Download image using curl
     subprocess.run("curl -o logo.jpg http://devops.witdemo.net/logo.jpg", shell=True)
 
-    # Create S3 client
-    s3 = boto3.client('s3')
+    s3Client = boto3.client("s3")
 
     # Upload image to S3 bucket and make it public
-    with open("logo.jpg", "rb") as f:
-        s3.upload_fileobj(f, bucketName, "logo.jpg", ExtraArgs={'ACL':'public-read', 'ContentType':'image/jpeg'})
-
-    # Enable static website hosting on the bucket
-    s3.put_bucket_website(
-        Bucket=bucketName,
-        WebsiteConfiguration={
-            'ErrorDocument': {'Key': 'error.html'},
-            'IndexDocument': {'Suffix': 'index.html'}
-        }
-    )
+    s3Client.upload_file("logo.jpg", bucket.name, "logo.jpg", ExtraArgs={'ACL':'public-read', 'ContentType':'image/jpeg'})
 
     # Create index.html file with image in the S3 bucket
     with open("index.html", "w") as f:
-        f.write(f'<html><body><img src="http://{bucketName}.s3.amazonaws.com/logo.jpg"></body></html>')
+        f.write(f'<html><body><img src="http://{bucket.name}.s3.amazonaws.com/logo.jpg"></body></html>')
 
     # Upload index.html to S3 bucket
-    with open("index.html", "rb") as f:
-        s3.upload_fileobj(f, bucketName, "index.html", ExtraArgs={'ACL':'public-read', 'ContentType':'text/html'})
+    s3Client.upload_file("index.html", bucket.name, "index.html", ExtraArgs={'ACL':'public-read', 'ContentType':'text/html'})
 
     # Opening the S3 bucket website in the browser
-    webbrowser.open_new_tab(f"http://{bucketName}.s3-website-us-east-1.amazonaws.com")
+    webbrowser.open_new_tab(f"http://{bucket.name}.s3-website-us-east-1.amazonaws.com")
     
     monitorCommands = f"""
-        scp -i {keyPairFilename} monitor.sh ec2-user@{instance.public_ip_address}:.
-        ssh -i {keyPairFilename} ec2-user@{instance.public_ip_address} 'chmod 700 monitor.sh'
-        ssh -i {keyPairFilename} ec2-user@{instance.public_ip_address} './monitor.sh'
+        scp -i {keyFilename} monitor.sh ec2-user@{instance.public_ip_address}:.
+        ssh -i {keyFilename} ec2-user@{instance.public_ip_address} 'chmod 700 monitor.sh'
+        ssh -i {keyFilename} ec2-user@{instance.public_ip_address} './monitor.sh'
     """
 
     subprocess.run(monitorCommands, shell=True)
